+++ flex_gemm_ap/kernels/triton/spconv/sparse_submanifold_conv_bwd_implicit_gemm_splitk.py
@@ -50,7 +50,7 @@ def sparse_submanifold_conv_bwd_input_implicit_gemm_splitk_kernel(
     num_k = tl.cdiv(Co, BK)  # Number of blocks in K dimension
     k_start = tl.cdiv(num_k * V * block_id_k, SPLITK)
     k_end = tl.cdiv(num_k * V * (block_id_k + 1), SPLITK)
-    offset_n = (block_id_n * B1 + tl.arange(0, B1)) % N         # (B1,)
+    offset_n = (block_id_n.to(tl.int64) * B1 + tl.arange(0, B1)) % N         # (B1,)
     offset_ci = (block_id_ci * B2 + tl.arange(0, B2)) % Ci      # (B2,)
     offset_k = tl.arange(0, BK)                                 # (BK,)
     
@@ -62,7 +62,7 @@ def sparse_submanifold_conv_bwd_input_implicit_gemm_splitk_kernel(
         v = k // num_k
         bk = k % num_k
         # Calculate pointers to grad_output matrix.
-        neighbor_offset_n = tl.load(neighbor + offset_n * V + V - 1 - v)                                    # (B1,)
+        neighbor_offset_n = tl.load(neighbor + offset_n.to(tl.int64) * V + V - 1 - v)                                    # (B1,)
         grad_output_ptr = grad_output + bk * BK + (neighbor_offset_n[:, None].to(tl.int64) * Co + offset_k[None, :])     # (B1, BK)
         # Calculate pointers to weight matrix.
         weight_ptr = weight + (((offset_k[:, None] + bk * BK) * V + v) * Ci + offset_ci[None, :])           # (BK, B2)
@@ -76,9 +76,9 @@ def sparse_submanifold_conv_bwd_input_implicit_gemm_splitk_kernel(
                              input_precision='tf32' if allow_tf32 else 'ieee')                              # (B1, B2)
                 
     # Write back the block of the output matrix with masks.
-    grad_input_offset_n = block_id_n * B1 + tl.arange(0, B1)
+    grad_input_offset_n = block_id_n.to(tl.int64) * B1 + tl.arange(0, B1)
     grad_input_offset_ci = block_id_ci * B2 + tl.arange(0, B2)
-    grad_input_ptr = grad_input + block_id_k * N * Ci + (grad_input_offset_n[:, None] * Ci + grad_input_offset_ci[None, :])
+    grad_input_ptr = grad_input + block_id_k.to(tl.int64) * N * Ci + (grad_input_offset_n[:, None].to(tl.int64) * Ci + grad_input_offset_ci[None, :])
     grad_input_mask = (grad_input_offset_n[:, None] < N) & (grad_input_offset_ci[None, :] < Ci)
     tl.store(grad_input_ptr, accumulator, mask=grad_input_mask)
 
@@ -132,8 +132,8 @@ def sparse_submanifold_conv_bwd_weight_implicit_gemm_splitk_kernel(
     offset_v = (tl.arange(0, BV) + (block_id_vci // (Ci // BCi)) * BV) % V          # (BV,)
     offset_ci = (tl.arange(0, BCi) + (block_id_vci % (Ci // BCi)) * BCi) % Ci       # (BCi,)
     offset_k = tl.arange(0, BK)                                                     # (BK,)
-    neighbor_ptr = neighbor + k_start * BK * V + (offset_k[:, None] * V + offset_v[None, :])            # (BK, BV)
-    grad_output_ptr = grad_output + k_start * BK * Co + (offset_k[None, :] * Co + offset_co[:, None])   # (B1, BK)
+    neighbor_ptr = neighbor + k_start.to(tl.int64) * BK * V + (offset_k[:, None].to(tl.int64) * V + offset_v[None, :])            # (BK, BV)
+    grad_output_ptr = grad_output + k_start.to(tl.int64) * BK * Co + (offset_k[None, :].to(tl.int64) * Co + offset_co[:, None])   # (B1, BK)
     
     # Create a block of the output matrix C.
     accumulator = tl.zeros((B1, BV * BCi), dtype=tl.float32)    
@@ -157,7 +157,7 @@ def sparse_submanifold_conv_bwd_weight_implicit_gemm_splitk_kernel(
     # Write back the block of the output matrix with masks.
     grad_weight_offset_co = block_id_co * B1 + tl.arange(0, B1)
     grad_weight_offset_vci = block_id_vci * BV * BCi + tl.arange(0, BV * BCi)
-    grad_weight_ptr = grad_weight + block_id_k * Co * V * Ci + (grad_weight_offset_co[:, None] * V * Ci + grad_weight_offset_vci[None, :])
+    grad_weight_ptr = grad_weight + block_id_k.to(tl.int64) * Co * V * Ci + (grad_weight_offset_co[:, None] * V * Ci + grad_weight_offset_vci[None, :])
     grad_weight_mask = (grad_weight_offset_co[:, None] < Co) & (grad_weight_offset_vci[None, :] < V * Ci)
     tl.store(grad_weight_ptr, accumulator, mask=grad_weight_mask)
